requirements.txt
# Core Machine Learning
torch==2.2.1
torchvision==0.17.1

# Utilities for Inference
Pillow==10.2.0
numpy==1.26.4

# Web Server
Flask==3.0.2
gunicorn==21.2.0

# pytorch-lightning==2.2.1